Add useWidgetFiltersQuery hook
Also, add the "reference" endpoint to the playground.

JIRA: RAIL-3383
