+++ libs/sdk-ui-dashboard/src/presentation/widget/common/index.ts
@@ -1,0 +1,2 @@
+// (C) 2020-2021 GoodData Corporation
+export * from "./useWidgetFiltersQuery";
